requirements.txt
streamlit
transformers
pillow
torch
sentencepiece